requirements.txt
streamlit
llama-index
llama-index-llms-groq
llama-index-embeddings-huggingface
python-dotenv
PyPDF2
pydantic